assignment 1

+++ assignment1/cs231n/features.py
@@ -108,6 +108,7 @@ def hog_feature(im):
   n_cellsy = int(np.floor(sy / cy))  # number of cells in y
   # compute orientations integral images
   orientation_histogram = np.zeros((n_cellsx, n_cellsy, orientations))
+
   for i in range(orientations):
     # create new integral image for this orientation
     # isolate orientations in this range
@@ -118,8 +119,8 @@ def hog_feature(im):
     # select magnitudes for those orientations
     cond2 = temp_ori > 0
     temp_mag = np.where(cond2, grad_mag, 0)
-    orientation_histogram[:,:,i] = uniform_filter(temp_mag, size=(cx, cy))[cx/2::cx, cy/2::cy].T
-  
+    orientation_histogram[:, :, i] = uniform_filter(temp_mag, size=(cx, cy))[int(cx / 2)::cx, int(cy / 2)::cy].T
+
   return orientation_histogram.ravel()
